Added addCustomerServiceRepForm, having an issue entering the data in the database. The error is talking specifically about the foreign that it cannot add the cID; my customer database is empty so I'm assuming that once I have data in it and I use a cID that exist then it will work.
Also, created a package name view and moved related class to it that was
created in the previous commit.

Additionally, I added a password attribute to the customerServiceRep
class and used it where necessary to get rid of an error (it's excepted
since it's old code). When everything is being merged the updated
CustomerServiceRep class will be used.

Next commit should have the addAccountsForm and addComplaintsForm
classes.